Add a test for concurrent use of global loggers
In working on UberFX, I realized that zap's global loggers are actually safe to
replace at runtime; because they're pointers, they're a single machine word.
Since loggers are also never mutated, we don't need any locks to make this code
race-free.

This commit removes the incorrect (and scary) documentation about races and adds
a test that attempts to trigger a race.