Add support for pkg/errors
Users who take the time (and performance hit) of using Dave Cheney's errors
package should get rich, annotated errors and stacktraces in their logs.
However, that package is only one of a number of similar packages hasn't yet
committed to a stable API, so we don't want to take a direct dependency on it.

To support these sorts of packages, serialize errors that implement
`fmt.Formatter` differently from those that don't.